+++ auth.js
@@ -117,4 +117,4 @@ function auth (functions, opts) {
   }
 }
 
-module.exports = fp(checkAuth, '>=3.0.0-alpha.1')
+module.exports = fp(checkAuth, '>=3')
